+++ Data/RunState.swift
@@ -6,4 +6,6 @@ import AppKit
  */
 class RunState {
     var color = NSColor.green
+    var filterRequired : [NSRegularExpression] = []
+    var filterExcluded : [NSRegularExpression] = []
 }