+++ LogLlamaTests/ScriptParserTest.swift
@@ -27,9 +27,6 @@ class ScriptParserTest: XCTestCase {
         assertThat(result, equalTo(true))
         assertThat(commands.count, equalTo(1))
         assertThat(commands[0], instanceOf(ReadFileCommand.self))
-        
-        let cmd = commands[0] as! ReadFileCommand
-        assertThat(cmd.file, equalTo("foo.txt"))
     }
     
     func test_parse_demo_recognizesIt() {
@@ -146,6 +143,63 @@ class ScriptParserTest: XCTestCase {
         assertThat(commands[0], instanceOf(TodayCommand.self))
     }
 
+    func test_parse_loadFilterRequired_recognizesIt() {
+        let sut = givenScriptParser()
+        let script = "require FNORD FJORD"
+        let (result, commands) = sut.parse(script: script)
+
+        assertThat(result, equalTo(true))
+        assertThat(commands.count, equalTo(1))
+        assertThat(commands[0], instanceOf(LoadFilterCommand.self))
+
+        let cmd = commands[0] as! LoadFilterCommand
+        _ = cmd.validate()
+        assertThat(cmd.loadFilterType, equalTo(LoadFilterCommand.LoadFilterType.Required))
+        assertThat(cmd.pattern, equalTo("FNORD FJORD"))
+    }
+
+    func test_parse_loadFilterExcluded_recognizesIt() {
+        let sut = givenScriptParser()
+        let script = "exclude FNORD FJORD"
+        let (result, commands) = sut.parse(script: script)
+
+        assertThat(result, equalTo(true))
+        assertThat(commands.count, equalTo(1))
+        assertThat(commands[0], instanceOf(LoadFilterCommand.self))
+
+        let cmd = commands[0] as! LoadFilterCommand
+        _ = cmd.validate()
+        assertThat(cmd.loadFilterType, equalTo(LoadFilterCommand.LoadFilterType.Excluded))
+        assertThat(cmd.pattern, equalTo("FNORD FJORD"))
+    }
+
+    func test_parse_loadFilterClear_recognizesIt() {
+        let sut = givenScriptParser()
+        let script = "clearFilters"
+        let (result, commands) = sut.parse(script: script)
+
+        assertThat(result, equalTo(true))
+        assertThat(commands.count, equalTo(1))
+        assertThat(commands[0], instanceOf(LoadFilterCommand.self))
+
+        let cmd = commands[0] as! LoadFilterCommand
+        _ = cmd.validate()
+        assertThat(cmd.loadFilterType, equalTo(LoadFilterCommand.LoadFilterType.Clear))
+    }
+
+    func test_parse_loadFilterRequireToday_recognizesIt() {
+        let sut = givenScriptParser()
+        let script = "requireToday"
+        let (result, commands) = sut.parse(script: script)
+
+        assertThat(result, equalTo(true))
+        assertThat(commands.count, equalTo(1))
+        assertThat(commands[0], instanceOf(LoadFilterCommand.self))
+
+        let cmd = commands[0] as! LoadFilterCommand
+        _ = cmd.validate()
+        assertThat(cmd.loadFilterType, equalTo(LoadFilterCommand.LoadFilterType.RequireToday))
+    }