+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased] 
 - Show current script file name in window title
-- Add support for glob patterns when loading log files with "<"
+- Add support for glob patterns when loading log files with `<`
+- add `require`, `exclude`, `requireToday`, and `clearFilter` commands to pre-filter lines as they are loaded.
 
 ## [1.0.0] - 2019-10-15
 - First release